+++ apps/client/src/app/modules/list/list.module.ts
@@ -57,6 +57,8 @@ import { ItemSourcesDisplayComponent } from './item/item-sources-display/item-so
 import { AlarmButtonComponent } from './item/alarm-button/alarm-button.component';
 import { ItemRowButtonsComponent } from './item/item-row-buttons/item-row-buttons.component';
 import { ListProgressbarComponent } from './list-progressbar/list-progressbar.component';
+import { ScrollingModule } from '@angular/cdk/scrolling';
+import { LazyScrollModule } from '../lazy-scroll/lazy-scroll.module';
 
 
 export const DATA_EXTRACTORS: Provider[] = [
@@ -112,7 +114,8 @@ export const DATA_EXTRACTORS: Provider[] = [
     ItemDetailsPopupsModule,
 
     StoreModule.forFeature('lists', listsReducer, { initialState: listsInitialState }),
-    EffectsModule.forFeature([ListsEffects])
+    EffectsModule.forFeature([ListsEffects]),
+    LazyScrollModule
   ],
   providers: [
     ...DATA_EXTRACTORS,