+++ apps/client/src/app/pages/inventory-optimizer/inventory-optimizer.module.ts
@@ -22,6 +22,8 @@ import { HasTooFew } from './optimizations/has-too-few';
 import { LazyDataService } from '../../core/data/lazy-data.service';
 import { ConsolidateStacks } from './optimizations/consolidate-stacks';
 import { UnwantedMaterials } from './optimizations/unwanted-materials';
+import { ScrollingModule } from '@angular/cdk/scrolling';
+import { LazyScrollModule } from '../../modules/lazy-scroll/lazy-scroll.module';
 
 const optimisations: Provider[] = [
   {
@@ -79,7 +81,9 @@ const routes: Routes = [
     PageLoaderModule,
     FullpageMessageModule,
     ItemIconModule,
-    ClipboardModule
+    ClipboardModule,
+    ScrollingModule,
+    LazyScrollModule
   ],
   providers: [
     ...optimisations